+++ apps/native-component-list/src/components/Page.tsx
@@ -1,9 +1,17 @@
 import { H4 } from '@expo/html-elements';
 import * as React from 'react';
-import { StyleSheet, View } from 'react-native';
+import { StyleSheet, View, ScrollView } from 'react-native';
 
 export function Page({ children }: { children: any }) {
   return <View style={{ paddingHorizontal: 12, paddingBottom: 12 }}>{children}</View>;
+}
+
+export function ScrollPage({ children }: { children: any }) {
+  return (
+    <ScrollView style={{ flex: 1, paddingHorizontal: 12, paddingBottom: 12 }}>
+      {children}
+    </ScrollView>
+  );
 }
 
 export function Section({ title, children, row }: { title: string; children: any; row?: boolean }) {